+++ pokete_classes/game.py
@@ -1,0 +1,1 @@
+game = None
